--- rlib-common/src/main/java/javasabr/rlib/common/plugin/extension/ExtensionPointManager.java
@@ -1,144 +1,0 @@
-package javasabr.rlib.common.plugin.extension;
-
-import javasabr.rlib.common.util.ClassUtils;
-import javasabr.rlib.common.util.dictionary.ConcurrentObjectDictionary;
-import javasabr.rlib.common.util.dictionary.DictionaryFactory;
-import javasabr.rlib.common.util.dictionary.ObjectDictionary;
-import javasabr.rlib.logger.api.Logger;
-import javasabr.rlib.logger.api.LoggerManager;
-import org.jspecify.annotations.NullMarked;
-
-/**
- * The manager to manage all extension points.
- *
- * @author JavaSaBr
- */
-@NullMarked
-public class ExtensionPointManager {
-
-  private static final Logger LOGGER = LoggerManager.getLogger(ExtensionPointManager.class);
-
-  /**
-   * Register a new extension point.
-   *
-   * @param id the extension id.
-   * @param <T> the extension's type.
-   * @return the new extension point.
-   * @throws IllegalArgumentException if an extension with the id is already exists.
-   */
-  public static <T> ExtensionPoint<T> register(String id) {
-    return getInstance().create(id);
-  }
-
-  private static final ExtensionPointManager INSTANCE = new ExtensionPointManager();
-
-  public static ExtensionPointManager getInstance() {
-    return INSTANCE;
-  }
-
-  /**
-   * All created extension points.
-   */
-  private final ConcurrentObjectDictionary<String, ExtensionPoint<?>> extensionPoints;
-
-  private ExtensionPointManager() {
-    this.extensionPoints = DictionaryFactory.newConcurrentAtomicObjectDictionary();
-  }
-
-  private <T> ExtensionPoint<T> create(String id) {
-
-    long stamp = extensionPoints.writeLock();
-    try {
-
-      var exists = extensionPoints.get(id);
-
-      if (exists != null) {
-        LOGGER.warning("The extension point with the id " + id + " is already registered.");
-        return ClassUtils.unsafeNNCast(exists);
-      }
-
-      var extensionPoint = new ExtensionPoint<T>();
-
-      extensionPoints.put(id, extensionPoint);
-
-      return extensionPoint;
-
-    } finally {
-      extensionPoints.writeUnlock(stamp);
-    }
-  }
-
-  /**
-   * Add the new extension to the extension point.
-   *
-   * @param id the extension point's id.
-   * @param type the extension's type.
-   * @param extension the new extension.
-   * @param <T> the extension's type.
-   * @return this manager.
-   */
-  public <T> ExtensionPointManager addExtension(String id, Class<T> type, T extension) {
-    getExtensionPoint(id).register(extension);
-    return this;
-  }
-
-  /**
-   * Add the new extension to the extension point.
-   *
-   * @param id the extension point's id.
-   * @param extension the new extension.
-   * @param <T> the extension's type.
-   * @return this manager.
-   */
-  public <T> ExtensionPointManager addExtension(String id, T extension) {
-    getExtensionPoint(id).register(extension);
-    return this;
-  }
-
-  /**
-   * Add the new extensions to the extension point.
-   *
-   * @param id the extension point's id.
-   * @param extensions the new extensions.
-   * @param <T> the extension's type.
-   * @return this manager.
-   */
-  public <T> ExtensionPointManager addExtension(String id, T... extensions) {
-    getExtensionPoint(id).register(extensions);
-    return this;
-  }
-
-  /**
-   * Get or create an extension point.
-   *
-   * @param id the id.
-   * @param type the extension's type.
-   * @param <T> the extension's type.
-   * @return the extension point.
-   */
-  public <T> ExtensionPoint<T> getExtensionPoint(String id, Class<T> type) {
-    return getExtensionPoint(id);
-  }
-
-  /**
-   * Get or create an extension point.
-   *
-   * @param id the id.
-   * @param <T> the extension's type.
-   * @return the extension point.
-   */
-  public <T> ExtensionPoint<T> getExtensionPoint(String id) {
-
-    var extensionPoint = extensionPoints.getInReadLock(id, ObjectDictionary::get);
-
-    if (extensionPoint != null) {
-      return ClassUtils.unsafeNNCast(extensionPoint);
-    }
-
-    try {
-      return create(id);
-    } catch (IllegalArgumentException e) {
-      return getExtensionPoint(id);
-    }
-  }
-}
