+++ rlib-common/src/test/java/javasabr/rlib/common/util/dictionary/ConcurrentObjectDictionaryTest.java
@@ -97,7 +97,7 @@ public class ConcurrentObjectDictionaryTest extends BaseTest {
       dic.put("3", 3);
     });
 
-    Integer val1 = dictionary.getInReadLock(
+    Integer val1 = dictionary.getFromReadLock(
         "1", (dic, arg) -> {
           assertType(dic, ConcurrentObjectDictionary.class);
           assertType(arg, String.class);
