+++ rlib-plugin-system/build.gradle
@@ -1,0 +1,4 @@
+
+dependencies {
+  api projects.rlibCommon
+}
