+++ rlib-plugin-system/src/main/java/javasabr/rlib/plugin/system/PluginSystem.java
@@ -1,0 +1,29 @@
+package javasabr.rlib.plugin.system;
+
+import java.util.Optional;
+import javasabr.rlib.common.util.array.Array;
+import org.jspecify.annotations.Nullable;
+
+/**
+ * @author JavaSaBr
+ */
+public interface PluginSystem {
+
+  Array<PluginContainer> pluginContainers();
+
+  @Nullable
+  PluginContainer getPluginContainer(String id);
+
+  default Optional<PluginContainer> getPluginContainerOptional(String id) {
+    return Optional.ofNullable(getPluginContainer(id));
+  }
+
+  Array<Plugin> plugins();
+
+  @Nullable
+  Plugin getPlugin(String id);
+
+  default Optional<Plugin> getPluginOptional(String id) {
+    return Optional.ofNullable(getPlugin(id));
+  }
+}
