+++ rlib-plugin-system/src/main/java/javasabr/rlib/plugin/system/annotation/PluginDescription.java
@@ -1,18 +1,16 @@
-package javasabr.rlib.common.plugin.annotation;
+package javasabr.rlib.plugin.system.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import org.jspecify.annotations.NullMarked;
 
 /**
  * The annotation to describe a plugin.
  *
  * @author JavaSaBr
  */
-@NullMarked
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE_USE)
