+++ rlib-plugin-system/src/main/java/javasabr/rlib/plugin/system/extension/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package javasabr.rlib.plugin.system.extension;
+
+import org.jspecify.annotations.NullMarked;
