+++ rlib-plugin-system/src/main/java/javasabr/rlib/plugin/system/impl/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package javasabr.rlib.plugin.system.impl;
+
+import org.jspecify.annotations.NullMarked;
