+++ rlib-plugin-system/src/main/java/javasabr/rlib/plugin/system/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package javasabr.rlib.plugin.system;
+
+import org.jspecify.annotations.NullMarked;
