+++ rlib-plugin-system/src/test/java/javasabr/rlib/plugin/system/PluginSystemTests.java
@@ -1,9 +1,8 @@
-package javasabr.rlib.common.plugin.system;
+package javasabr.rlib.plugin.system;
 
 import java.nio.file.Paths;
 import java.util.concurrent.ForkJoinPool;
-import javasabr.rlib.common.plugin.Version;
-import javasabr.rlib.common.plugin.impl.PluginSystemFactory;
+import javasabr.rlib.plugin.system.impl.PluginSystemFactory;
 import org.junit.jupiter.api.Test;
 
 /**
@@ -15,12 +14,13 @@ public class PluginSystemTests {
   public void test() {
 
     var pluginSystem = PluginSystemFactory.newBasePluginSystem();
-    pluginSystem.setAppVersion(new Version("0.0.1"));
+    pluginSystem.configureAppVersion(new Version("0.0.1"));
     pluginSystem.configureEmbeddedPluginPath(Paths.get("../gradle/"));
 
     pluginSystem
         .preLoad(ForkJoinPool.commonPool())
         .thenApply(system -> system.initialize(ForkJoinPool.commonPool()))
+        .toCompletableFuture()
         .join();
   }
 }
