+++ settings.gradle
@@ -10,3 +10,4 @@ include ':rlib-mail'
 include ':rlib-logger-api'
 include ':rlib-logger-impl'
 include ':rlib-logger-slf4j'
+include ':rlib-plugin-system'
